# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "FieldViolation": {
        "additionalProperties": false,
        "description": "One request-validation failure safe to expose to an API client.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldViolation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "description": "Cursor-free pagination metadata for the initial API contract.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "PageInfo",
        "type": "object"
      },
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "RFC 9457-style error response with a stable application code.",
        "examples": [
          {
            "code": "FEATURE_NOT_READY",
            "detail": "Document upload is not implemented yet.",
            "errors": [],
            "instance": "/api/v1/documents/upload",
            "status": 501,
            "title": "Feature not ready",
            "type": "urn:docqc:error:FEATURE_NOT_READY"
          }
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "StandardPatternCreate": {
        "additionalProperties": false,
        "description": "Admin request to register a normalized standards-code pattern.",
        "properties": {
          "code_family": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Code Family",
            "type": "string"
          },
          "code_pattern": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Code Pattern",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "organization_scope": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Organization Scope",
            "type": "string"
          }
        },
        "required": [
          "code_family",
          "code_pattern",
          "description",
          "organization_scope"
        ],
        "title": "StandardPatternCreate",
        "type": "object"
      },
      "StandardPatternListResponse": {
        "additionalProperties": false,
        "description": "Paginated standards-registry collection.",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "standards": {
            "items": {
              "$ref": "#/components/schemas/StandardPatternRead"
            },
            "title": "Standards",
            "type": "array"
          }
        },
        "required": [
          "standards",
          "pagination"
        ],
        "title": "StandardPatternListResponse",
        "type": "object"
      },
      "StandardPatternRead": {
        "additionalProperties": false,
        "description": "Persisted standards-registry entry.",
        "properties": {
          "code_family": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Code Family",
            "type": "string"
          },
          "code_pattern": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Code Pattern",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "organization_scope": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Organization Scope",
            "type": "string"
          }
        },
        "required": [
          "code_family",
          "code_pattern",
          "description",
          "organization_scope",
          "id",
          "created_at"
        ],
        "title": "StandardPatternRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Automated first-pass quality review of engineering documents (PDF/DOCX)",
    "title": "Document QC API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/standards-registry": {
      "get": {
        "description": "List configured standards-code patterns.",
        "operationId": "list_standard_patterns_api_v1_standards_registry_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardPatternListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "List Standard Patterns",
        "tags": [
          "standards"
        ]
      },
      "post": {
        "description": "Register an organization-scoped standards-code pattern.",
        "operationId": "create_standard_pattern_api_v1_standards_registry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardPatternCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardPatternRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Create Standard Pattern",
        "tags": [
          "standards"
        ]
      }
    },
...
  }
}